+++ message_ix/tests/test_report.py
@@ -9,6 +9,8 @@ import pandas as pd
 import pyam
 import pytest
 from genno.testing import assert_qty_equal
+from ixmp import Platform
+from ixmp.backend.jdbc import JDBCBackend
 from ixmp.report import Reporter as ixmp_Reporter
 from ixmp.testing import assert_logs
 from numpy.testing import assert_allclose
@@ -18,9 +20,14 @@ from message_ix import Scenario
 from message_ix.report import Reporter, configure
 from message_ix.testing import SCENARIO, make_dantzig, make_westeros
 
+# NOTE These tests maybe don't need to be parametrized.
+# Does `Reporter.from_scenario()` depend on otherwise untested Scenario functions?
+
 
 class TestReporter:
-    def test_add_sankey(self, test_mp, request) -> None:
+    def test_add_sankey(
+        self, test_mp: Platform, request: pytest.FixtureRequest
+    ) -> None:
         scen = make_westeros(test_mp, solve=True, quiet=True, request=request)
         rep = Reporter.from_scenario(scen, units={"replace": {"-": ""}})
 
@@ -33,7 +40,9 @@ class TestReporter:
         assert rep.check_keys(key)
 
 
-def test_reporter_no_solution(caplog, message_test_mp) -> None:
+def test_reporter_no_solution(
+    caplog: pytest.LogCaptureFixture, message_test_mp: Platform
+) -> None:
     scen = Scenario(message_test_mp, **SCENARIO["dantzig"])
 
     with assert_logs(
@@ -51,7 +60,14 @@ def test_reporter_no_solution(caplog, message_test_mp) -> None:
     assert 3 == len(result)
 
 
-def test_reporter_from_scenario(message_test_mp) -> None:
+# IXMP4Backend is currently not storing the MACRO variables 'C' and 'I' for MESSAGE
+# models.
+MISSING_IXMP4 = {"C:", "C:n", "C:n-y", "C:y", "I:", "I:n", "I:n-y", "I:y"}
+
+
+def test_reporter_from_scenario(
+    message_test_mp: Platform, test_data_path: Path
+) -> None:
     scen = Scenario(message_test_mp, **SCENARIO["dantzig"])
 
     # Varies between local & CI contexts
@@ -65,8 +81,10 @@ def test_reporter_from_scenario(message_test_mp) -> None:
     # message_ix.Reporter can also be initialized
     rep = Reporter.from_scenario(scen)
 
-    # Number of quantities available in a rudimentary MESSAGEix Scenario
-    assert 268 == len(rep.graph["all"])
+    # NOTE Used to write out the expected data
+    # Path(test_data_path / "reportergraph.txt").write_text(
+    #     "\n".join(list(map(str, sorted(rep.graph))))
+    # )
 
     # Quantities have short dimension names
     assert "demand:n-c-l-y-h" in rep, sorted(rep.graph)
@@ -83,12 +101,22 @@ def test_reporter_from_scenario(message_test_mp) -> None:
     # check_attrs False because we don't get the unit addition in bare xarray
     assert_qty_equal(obs, demand, check_attrs=False)
 
+    # Prepare the expected items in the graphs
+    expected_rep_ix_graph_keys = set(
+        Path(test_data_path / "reporterixgraph.txt").read_text().split("\n")
+    )
+    expected_rep_graph_keys = set(
+        Path(test_data_path / "reportergraph.txt").read_text().split("\n")
+    )
+    if not isinstance(message_test_mp._backend, JDBCBackend):
+        expected_rep_ix_graph_keys -= MISSING_IXMP4
+        expected_rep_graph_keys -= MISSING_IXMP4
+
     # ixmp.Reporter pre-populated with only model quantities and aggregates
-    assert 6477 == len(rep_ix.graph)
+    assert set(map(str, sorted(rep_ix.graph))) == expected_rep_ix_graph_keys
 
     # message_ix.Reporter pre-populated with additional, derived quantities
-    # This is the same value as in test_tutorials.py
-    assert 13739 == len(rep.graph)
+    assert set(map(str, sorted(rep.graph))) == expected_rep_graph_keys
 
     # Derived quantities have expected dimensions
     vom_key = rep.full_key("vom")
@@ -103,7 +131,9 @@ def test_reporter_from_scenario(message_test_mp) -> None:
     assert_qty_equal(vom, rep.get(vom_key))
 
 
-def test_reporter_from_dantzig(request, test_mp) -> None:
+def test_reporter_from_dantzig(
+    request: pytest.FixtureRequest, test_mp: Platform
+) -> None:
     scen = make_dantzig(test_mp, solve=True, quiet=True, request=request)
 
     # Reporter.from_scenario can handle Dantzig example model
@@ -113,7 +143,9 @@ def test_reporter_from_dantzig(request, test_mp) -> None:
     rep.get("all")
 
 
-def test_reporter_from_westeros(request, test_mp) -> None:
+def test_reporter_from_westeros(
+    request: pytest.FixtureRequest, test_mp: Platform
+) -> None:
     scen = make_westeros(test_mp, emissions=True, solve=True, request=request)
 
     # Reporter.from_scenario can handle Westeros example model
@@ -154,7 +186,9 @@ def test_reporter_from_westeros(request, test_mp) -> None:
     assert_allclose(df["value_y"], df["value_x"], err_msg=df.to_string())
 
 
-def test_reporter_as_pyam(caplog, tmp_path, dantzig_reporter) -> None:
+def test_reporter_as_pyam(
+    caplog: pytest.LogCaptureFixture, tmp_path: Path, dantzig_reporter: Reporter
+) -> None:
     caplog.set_level(logging.INFO)
 
     rep = dantzig_reporter
@@ -167,6 +201,7 @@ def test_reporter_as_pyam(caplog, tmp_path, dantzig_reporter) -> None:
     rename = dict(nl="region", ya="year")
 
     # Add a task that converts ACT to a pyam.IamDataFrame
+    assert as_pyam
     rep.add("ACT IAMC", (partial(as_pyam, rename=rename, drop=["yv"]), "scenario", ACT))
 
     # Result is an IamDataFrame
@@ -195,6 +230,7 @@ def test_reporter_as_pyam(caplog, tmp_path, dantzig_reporter) -> None:
     key2 = rep.add("as_pyam", ACT, "iamc", rename=rename, collapse=add_tm)
 
     # Keys of added node(s) are returned
+    assert isinstance(ACT, genno.Key)
     assert ACT.name + "::iamc" == key2
 
     caplog.clear()