+++ message_ix/tests/tools/test_migrate.py
@@ -3,17 +3,21 @@
 import sys
 
 import pytest
+from ixmp import Platform
 
 from message_ix.testing import make_westeros
 from message_ix.tools.migrate import v311
 
 
+@pytest.mark.jdbc
 @pytest.mark.xfail(
     condition=sys.version_info < (3, 10),
     raises=ImportError,
     reason="Uses itertools.pairwise, added in Python 3.10",
 )
-def test_v311(caplog, request, test_mp) -> None:
+def test_v311(
+    caplog: pytest.LogCaptureFixture, request: pytest.FixtureRequest, test_mp: Platform
+) -> None:
     """Minimal test of :func:`.v311`."""
     s = make_westeros(test_mp, request=request)